+++ src/core/Flora/Import/Package.hs
@@ -352,6 +352,7 @@ extractPackageDataFromCabal userId (repositoryName, repositoryPackages) uploadTi
           , changelog = Nothing
           , changelogStatus = NotImported
           , repository = Just repositoryName
+          , tarball = Nothing
           , license = Cabal.license packageDesc
           , sourceRepos
           , homepage = Just $ display packageDesc.homepage