+++ src/jobs-worker/FloraJobs/ThirdParties/Hackage/Client.hs
@@ -4,11 +4,13 @@
 module FloraJobs.ThirdParties.Hackage.Client where
 
 import Control.Monad.IO.Class
+import Data.ByteString.Lazy (ByteString)
 import Data.Proxy
 import Data.Text
 import Data.Time (UTCTime)
 import Data.Time.Orphans ()
 import Data.Vector (Vector)
+import Effectful (Eff, IOE, type (:>))
 import Effectful.Reader.Static
 import Network.HTTP.Req (GET (GET), NoReqBody (..))
 import Network.HTTP.Req qualified as Req
@@ -17,9 +19,9 @@ import Servant.Client (BaseUrl (..), Client, ClientError (..), ClientM, Scheme (
 
 import Flora.Model.Package.Types
 import FloraJobs.ThirdParties.Hackage.API as API
-import FloraJobs.Types (JobsRunner, JobsRunnerEnv (..))
+import FloraJobs.Types (JobsRunnerEnv (..))
 
-request :: ClientM a -> JobsRunner (Either ClientError a)
+request :: (IOE :> es, Reader JobsRunnerEnv :> es) => ClientM a -> Eff es (Either ClientError a)
 request req = do
   JobsRunnerEnv{httpManager} <- ask
   let clientEnv =
@@ -36,6 +38,14 @@ listHackageUsers = hackageClient // API.listUsers
 
 getHackageUser :: Text -> ClientM HackageUserDetailsObject
 getHackageUser username = hackageClient // API.withUser /: username // API.getUser
+
+getPackageTarball :: VersionedPackage -> ClientM ByteString
+getPackageTarball versionedPackage =
+  hackageClient
+    // API.withPackage
+    /: versionedPackage
+    // API.getTarball
+    /: VersionedTarball versionedPackage
 
 getPackageReadme :: VersionedPackage -> ClientM Text
 getPackageReadme versionedPackage =